auto goland refactor ot tests/helpers.go to internal/testutil/testutil.go
