+++ cardinal/ecs/internal/testutil/testutil.go
@@ -1,4 +1,4 @@
-package tests
+package testutil
 
 import (
 	"context"
