+++ cardinal/ecs/storage/nonce_test.go
@@ -2,15 +2,14 @@ package storage_test
 
 import (
 	"fmt"
+	"pkg.world.dev/world-engine/cardinal/ecs/internal/testutil"
 	"testing"
 
 	"gotest.tools/v3/assert"
-
-	"pkg.world.dev/world-engine/cardinal/ecs/tests"
 )
 
 func TestSetAndGetNonce(t *testing.T) {
-	rs := tests.GetRedisStorage(t)
+	rs := testutil.GetRedisStorage(t)
 	address := "some-address"
 	wantNonce := uint64(100)
 	assert.NilError(t, rs.SetNonce(address, wantNonce))
@@ -20,7 +19,7 @@ func TestSetAndGetNonce(t *testing.T) {
 }
 
 func TestMissingNonceIsZero(t *testing.T) {
-	rs := tests.GetRedisStorage(t)
+	rs := testutil.GetRedisStorage(t)
 
 	gotNonce, err := rs.GetNonce("some-address-that-doesn't-exist")
 	assert.NilError(t, err)
@@ -28,7 +27,7 @@ func TestMissingNonceIsZero(t *testing.T) {
 }
 
 func TestCanStoreManyNonces(t *testing.T) {
-	rs := tests.GetRedisStorage(t)
+	rs := testutil.GetRedisStorage(t)
 	for i := uint64(10); i < 100; i++ {
 		addr := fmt.Sprintf("%d", i)
 		assert.NilError(t, rs.SetNonce(addr, i))
