+++ cardinal/ecs/tick_test.go
@@ -3,6 +3,7 @@ package ecs_test
 import (
 	"context"
 	"errors"
+	"pkg.world.dev/world-engine/cardinal/ecs/internal/testutil"
 	"testing"
 
 	"github.com/alicebob/miniredis/v2"
@@ -11,12 +12,11 @@ import (
 	"pkg.world.dev/world-engine/cardinal/ecs"
 	"pkg.world.dev/world-engine/cardinal/ecs/inmem"
 	"pkg.world.dev/world-engine/cardinal/ecs/storage"
-	"pkg.world.dev/world-engine/cardinal/ecs/tests"
 )
 
 func TestTickHappyPath(t *testing.T) {
 	rs := miniredis.RunT(t)
-	oneWorld := tests.InitWorldWithRedis(t, rs)
+	oneWorld := testutil.InitWorldWithRedis(t, rs)
 	oneEnergy := ecs.NewComponentType[EnergyComponent]()
 	assert.NilError(t, oneWorld.RegisterComponents(oneEnergy))
 	assert.NilError(t, oneWorld.LoadGameState())
@@ -27,7 +27,7 @@ func TestTickHappyPath(t *testing.T) {
 
 	assert.Equal(t, uint64(10), oneWorld.CurrentTick())
 
-	twoWorld := tests.InitWorldWithRedis(t, rs)
+	twoWorld := testutil.InitWorldWithRedis(t, rs)
 	twoEnergy := ecs.NewComponentType[EnergyComponent]()
 	assert.NilError(t, twoWorld.RegisterComponents(twoEnergy))
 	assert.NilError(t, twoWorld.LoadGameState())
@@ -40,7 +40,7 @@ func TestCanIdentifyAndFixSystemError(t *testing.T) {
 	}
 
 	rs := miniredis.RunT(t)
-	oneWorld := tests.InitWorldWithRedis(t, rs)
+	oneWorld := testutil.InitWorldWithRedis(t, rs)
 	onePower := ecs.NewComponentType[PowerComponent]()
 	assert.NilError(t, oneWorld.RegisterComponents(onePower))
 
@@ -71,7 +71,7 @@ func TestCanIdentifyAndFixSystemError(t *testing.T) {
 	assert.ErrorIs(t, errorSystem, oneWorld.Tick(context.Background()))
 
 	// Set up a new world using the same storage layer
-	twoWorld := tests.InitWorldWithRedis(t, rs)
+	twoWorld := testutil.InitWorldWithRedis(t, rs)
 	twoPower := ecs.NewComponentType[*PowerComponent]()
 	assert.NilError(t, twoWorld.RegisterComponents(twoPower))
 
@@ -146,7 +146,7 @@ func TestCanRecoverStateAfterFailedArchetypeChange(t *testing.T) {
 	}
 	rs := miniredis.RunT(t)
 	for _, firstWorldIteration := range []bool{true, false} {
-		world := tests.InitWorldWithRedis(t, rs)
+		world := testutil.InitWorldWithRedis(t, rs)
 		static := ecs.NewComponentType[ScalarComponent]()
 		toggle := ecs.NewComponentType[ScalarComponent]()
 		assert.NilError(t, world.RegisterComponents(static, toggle))
@@ -215,7 +215,7 @@ func TestCanRecoverTransactionsFromFailedSystemRun(t *testing.T) {
 	rs := miniredis.RunT(t)
 	errorBadPowerChange := errors.New("bad power change transaction")
 	for _, isBuggyIteration := range []bool{true, false} {
-		world := tests.InitWorldWithRedis(t, rs)
+		world := testutil.InitWorldWithRedis(t, rs)
 
 		powerComp := ecs.NewComponentType[FloatValue]()
 		assert.NilError(t, world.RegisterComponents(powerComp))
